REF: Refactor estimator index checks to integration test file
Refactor estimator index checks to integration test file.
