+++ test/test_estimator.py
@@ -21,31 +21,14 @@
 #     https://www.nipreps.org/community/licensing/
 #
 
-from typing import Union
-
 import numpy as np
-import pytest
 
-import nifreeze.estimator
 from nifreeze.data.base import BaseDataset
-from nifreeze.data.dmri.utils import DEFAULT_LOWB_THRESHOLD
-from nifreeze.data.pet.utils import compute_uptake_statistic
 from nifreeze.estimator import Estimator
 from nifreeze.model.base import BaseModel
 from nifreeze.utils import iterators
 
 DATAOBJ_SIZE = (5, 5, 5, 7)
-
-import re
-from nifreeze.utils.iterators import (
-    ITERATOR_MULTIPLICITY_ERROR_MSG,
-    ITERATOR_SIZE_ERROR_MSG,
-    START_INDEX_POSITIVITY_ERROR_MSG,
-    START_INDEX_DATA_LENGTH_ERROR_MSG,
-    STOP_INDEX_ORDERING_ERROR_MSG,
-    STOP_INDEX_DATA_LENGTH_ERROR_MSG,
-    SIZE_KWARG,
-)
 
 
 class DummyInsiderModel(BaseModel):
@@ -129,186 +112,3 @@ def test_estimator_init_model_string(request, monkeypatch):
     est.run(_dataset)
     assert isinstance(est._model, str)
     assert est._model == model_name
-
-
-def test_estimator_start_index(request, monkeypatch):
-    """Test that estimator respects start_index."""
-    recorded_indices = []
-
-    def fake_set_transform(self, i, xform):
-        recorded_indices.append(i)
-
-    monkeypatch.setattr(
-        type(DummyDataset(rng=request.node.rng)), "set_transform", fake_set_transform
-    )
-
-    class DummyXForm:
-        matrix = np.eye(4)
-
-    monkeypatch.setattr(
-        nifreeze.estimator,
-        "_run_registration",
-        lambda *a, **k: DummyXForm(),
-    )
-
-    dataset = DummyDataset(rng=request.node.rng)
-    model = DummyInsiderModel(dataset=dataset)
-
-    estimator = Estimator(model, strategy="linear", start_index=3)
-    estimator.run(dataset)
-
-    assert recorded_indices == [3, 4, 5, 6]
-
-
-def test_estimator_start_and_end_index(request, monkeypatch):
-    """Test that estimator respects both start_index and end_index."""
-    recorded_indices = []
-
-    def fake_set_transform(self, i, xform):
-        recorded_indices.append(i)
-
-    monkeypatch.setattr(
-        type(DummyDataset(rng=request.node.rng)), "set_transform", fake_set_transform
-    )
-
-    class DummyXForm:
-        matrix = np.eye(4)
-
-    monkeypatch.setattr(
-        nifreeze.estimator,
-        "_run_registration",
-        lambda *a, **k: DummyXForm(),
-    )
-
-    dataset = DummyDataset(rng=request.node.rng)
-    model = DummyInsiderModel(dataset=dataset)
-
-    estimator = Estimator(model, strategy="linear", start_index=2, stop_index=5)
-    estimator.run(dataset)
-
-    # Should only process indices 2-4 (stop_index is exclusive)
-    assert recorded_indices == [2, 3, 4]
-
-
-def test_estimator_invalid_start_index(request):
-    """Test that negative start_index raises an error."""
-    dataset = DummyDataset(rng=request.node.rng)
-    model = DummyInsiderModel(dataset=dataset)
-
-    estimator = Estimator(model, start_index=-1)
-    with pytest.raises(ValueError, match=re.escape(START_INDEX_POSITIVITY_ERROR_MSG)):
-        estimator.run(dataset)
-
-    estimator = Estimator(model, start_index=len(dataset))
-    with pytest.raises(
-        ValueError, match=re.escape(START_INDEX_DATA_LENGTH_ERROR_MSG.format(feature=SIZE_KWARG))
-    ):
-        estimator.run(dataset)
-
-
-def test_estimator_invalid_stop_index(request):
-    """Test that stop_index <= start_index raises an error."""
-    dataset = DummyDataset(rng=request.node.rng)
-    model = DummyInsiderModel(dataset=dataset)
-
-    estimator = Estimator(model, start_index=2, stop_index=2)
-    with pytest.raises(ValueError, match=re.escape(STOP_INDEX_ORDERING_ERROR_MSG)):
-        estimator.run(dataset)
-
-    estimator = Estimator(model, stop_index=len(dataset) + 1)
-    with pytest.raises(ValueError, match=re.escape(STOP_INDEX_DATA_LENGTH_ERROR_MSG.format(feature=SIZE_KWARG))):
-        estimator.run(dataset)
-
-
-@pytest.mark.parametrize(
-    "strategy, iterator_func, modality",
-    [
-        ("linear", iterators.linear_iterator, "dwi"),
-        ("linear", iterators.linear_iterator, "pet"),
-        ("random", iterators.random_iterator, "dwi"),
-        ("random", iterators.random_iterator, "pet"),
-        ("centralsym", iterators.centralsym_iterator, "dwi"),
-        ("centralsym", iterators.centralsym_iterator, "pet"),
-        ("monotonic_value", iterators.monotonic_value_iterator, "dwi"),
-        ("monotonic_value", iterators.monotonic_value_iterator, "pet"),
-    ],
-)
-def test_estimator_iterator_index_match(
-    monkeypatch, setup_random_dwi_data, setup_random_pet_data, strategy, iterator_func, modality
-):
-    dataset: Union["DummyDWIDataset", "DummyPETDataset"]  # Avoids type annotation errors
-    if modality == "dwi":
-        dwi_dataobj, affine, brainmask_dataobj, gradients, b0_thres = setup_random_dwi_data
-
-        b0s_mask = gradients[:, -1] <= b0_thres
-        b0_dataobj = dwi_dataobj[..., b0s_mask].squeeze()
-
-        dataset = DummyDWIDataset(dwi_dataobj, affine, brainmask_dataobj, b0_dataobj, gradients)
-        bvals = gradients[:, -1][gradients[:, -1] > DEFAULT_LOWB_THRESHOLD]
-        kwargs = {"bvals": bvals}
-    elif modality == "pet":
-        (
-            pet_dataobj,
-            affine,
-            brainmask_dataobj,
-            _,
-            midframe,
-            total_duration,
-        ) = setup_random_pet_data
-
-        dataset = DummyPETDataset(pet_dataobj, affine, brainmask_dataobj, midframe, total_duration)
-        uptake = compute_uptake_statistic(pet_dataobj, stat_func=np.sum)
-        kwargs = {"uptake": uptake}
-    else:
-        raise NotImplementedError(f"{modality} not implemented")
-
-    # Patch set_transform to record indices and matrices
-    recorded_indices = []
-    recorded_matrices = []
-
-    # Make this accept `self` so it behaves as a proper instance method
-    def fake_set_transform(self, i, xform):
-        recorded_indices.append(i)
-        recorded_matrices.append(xform)
-
-    monkeypatch.setattr(type(dataset), "set_transform", fake_set_transform)
-
-    # Patch registration to return identity matrix
-    class DummyXForm:
-        matrix = np.eye(4)
-
-    monkeypatch.setattr(
-        nifreeze.estimator,
-        "_run_registration",
-        lambda *a, **k: DummyXForm(),
-    )
-
-    model = DummyInsiderModel(dataset=dataset)
-    estimator = Estimator(model, strategy=strategy)
-    estimator.run(dataset, **kwargs)
-
-    n_vols = len(dataset)
-
-    # Get expected indices
-    if strategy == "linear":
-        expected_indices = list(iterator_func(size=n_vols))
-    elif strategy == "random":
-        expected_indices = sorted(iterator_func(size=n_vols, seed=42))
-        recorded_indices_sorted = sorted(recorded_indices)
-        assert recorded_indices_sorted == expected_indices
-        return
-    elif strategy == "centralsym":
-        expected_indices = list(iterator_func(size=n_vols))
-    elif strategy == "monotonic_value":
-        if modality == "dwi":
-            expected_indices = list(iterator_func(bvals=bvals, ascending=True))
-        elif modality == "pet":
-            expected_indices = list(iterator_func(uptake=uptake, ascending=False))
-        else:
-            raise NotImplementedError(f"Modality {modality} not implemented")
-    else:
-        raise ValueError(f"Unknown strategy {strategy}")
-
-    # Assert indices and matrices
-    assert recorded_indices == expected_indices
-    assert all(np.allclose(mat, np.eye(4)) for mat in recorded_matrices)
